Initial checkin to make this into a pip package
fix
