+++ provision/.gitignore
@@ -1,2 +1,5 @@
 *~
 *.pyc
+/*.egg
+/*.egg-info
+/dist/
