+++ provision/MANIFEST.in
@@ -1,0 +1,3 @@
+include README.md
+include acc_provision/*.yaml
+include acc_provision/templates/*.yaml
