+++ provision/Makefile
@@ -2,7 +2,10 @@
 # Minimal Makefile to run aci-containers-provisioning tests
 #
 
+dist: test
+	python setup.py sdist
+
 test:
-	pep8 --ignore E501 aci-containers-provision.py apic.py
-	flake8 --ignore E501 aci-containers-provision.py apic.py
-	py.test aci-containers-provision.py apic.py
+	pep8 --ignore E501 acc_provision
+	flake8 --ignore E501 acc_provision
+	py.test acc_provision
