+++ provision/acc_provision/templates/aci-containers.yaml
